requirements.txt
PyYAML>=3.12
requests>=2.18.4